Refactor upload resource integration test
- consolidate JVM, web server and web application resource upload scenarios into one feature
- remove "I load properties file"
- maintain only one unit test to run JVM, web server and web application resource upload
