--- jwala-integration-test/src/test/java/com/cerner/jwala/ui/selenium/configuration/ManageWebappResourceTest.java
@@ -1,16 +1,0 @@
-package com.cerner.jwala.ui.selenium.configuration;
-
-import cucumber.api.CucumberOptions;
-import cucumber.api.junit.Cucumber;
-import org.junit.runner.RunWith;
-
-/**
- * Created by Rahul Sayini on 7/10/2017.
- */
-
-@RunWith(Cucumber.class)
-@CucumberOptions(
-        features = { "classpath:com/cerner/jwala/ui/selenium/configuration/uploadWebappResource.feature" },
-        glue = {"com.cerner.jwala.ui.selenium.steps"})
-public class ManageWebappResourceTest {
-}
