+++ jwala-integration-test/src/test/resources/selenium/parameters.properties
@@ -1,7 +1,8 @@
 implementation-version=
 host1=someHost1
 host2=someHost2
-media.remote.dir=c:/stp
+media.remote.dir=d:/ctp
 httpdDeployPath=d:/ctp/app/data/httpd/seleniumWebserver
 webAppDeployPath=d:/ctp/app/webapps
 resource.deploy.path=d:/ctp/app/instances
+httpd.resource.deploy.path=d:/ctp/app/data/httpd
